+++ src/signposting/htmllinks.py
@@ -51,7 +51,7 @@ class DownloadedText(str):
     """The requested URL, before redirection"""
 
     resolved_url: AbsoluteURI
-    """The resolved URL, after redirection and observing any Content-Location header."""
+    """The resolved URL, after redirection."""
 
     def __new__(cls, value:str, content_type:str, requested_url:AbsoluteURI, resolved_url:AbsoluteURI):
         # NOTE: Do not return value if it's already an DownloadedText
@@ -87,9 +87,14 @@ def _get_html(uri:AbsoluteURI) -> Union[HTML,XHTML]:
     page = requests.get(uri, headers=HEADERS)
 
     resolved_url = AbsoluteURI(page.url, uri)
-    if "Content-Location" in page.headers:
-        # More specific, e.g. "index.en.html" - parse as relative URI reference
-        resolved_url = AbsoluteURI(page.headers["Content-Location"], resolved_url)
+    
+    # Note: According to HTTP/1.1 updates (Appendix B) in 
+    # https://datatracker.ietf.org/doc/html/rfc7231
+    # then Content-Location should NO LONGER be used for 
+    # resolving relative URI references.
+    ##if "Content-Location" in page.headers:
+    ##    # More specific, e.g. "index.en.html" - parse as relative URI reference
+    ##    resolved_url = AbsoluteURI(page.headers["Content-Location"], resolved_url)
 
     if page.status_code == 203:
         warnings.warn("203 Non-Authoritative Information <%s> - Signposting URIs may have been rewritten by proxy" %