+++ tests/test_linkset.py
@@ -45,9 +45,14 @@ class TestParseLinkset(unittest.TestCase):
         ls = linkset.Linkset(a2a_28, "application/linkset", 
             AbsoluteURI("https://s11.example.net/2022/a2a-fair-metrics/28-http-linkset-txt-only/linkset.txt"),
             AbsoluteURI("https://s11.example.net/2022/a2a-fair-metrics/28-http-linkset-txt-only/linkset.txt"))
-        signposts = linkset._parse_linkset(ls)
-        self.assertEqual("https://s11.example.net/2022/a2a-fair-metrics/28-http-linkset-txt-only/linkset.txt", signposts.context_url)
-        
+        metasignposts = linkset._parse_linkset(ls)
+        self.assertEqual("https://s11.example.net/2022/a2a-fair-metrics/28-http-linkset-txt-only/linkset.txt", metasignposts.context_url)
+
+        ## Linksets, as separate HTTP resources, may contain links for multiple context,         
+        ## therefore we have to select the right one
+        self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/"}, metasignposts.other_contexts)
+        signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/")
+
         self.assertEqual("https://w3id.org/a2a-fair-metrics/28-http-linkset-txt-only/", signposts.citeAs.target)
 
         self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/index.ttl"},
@@ -69,9 +74,12 @@ class TestParseLinkset(unittest.TestCase):
         ls = linkset.LinksetJSON(a2a_27, "application/linkset+json", 
             AbsoluteURI("https://s11.example.net/2022/a2a-fair-metrics/27-http-linkset-json-only/linkset.json"),
             AbsoluteURI("https://s11.example.net/2022/a2a-fair-metrics/27-http-linkset-json-only/linkset.json"))
-        signposts = linkset._parse_linkset_json(ls)
-        self.assertEqual("https://s11.example.net/2022/a2a-fair-metrics/27-http-linkset-json-only/linkset.json", signposts.context_url)
-        
+        metasignposts = linkset._parse_linkset_json(ls)
+
+        self.assertEqual("https://s11.example.net/2022/a2a-fair-metrics/27-http-linkset-json-only/linkset.json", metasignposts.context_url)
+        self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/"}, metasignposts.other_contexts)
+        signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/")
+
         self.assertEqual("https://w3id.org/a2a-fair-metrics/27-http-linkset-json-only/", signposts.citeAs.target)
         
         self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/index.ttl"},
@@ -102,9 +110,11 @@ class TestFindSignpostingLinkset(unittest.TestCase):
     def test_parse_linkset_a2a_07(self):
         PID = "https://w3id.org/a2a-fair-metrics/07-http-describedby-citeas-linkset-json/"
         for (uri,_) in self._linksets_for_pid(PID):
-            signposts = linkset.find_signposting_linkset(uri)
-            self.assertEqual(uri, signposts.context_url)
-            
+            metasignposts = linkset.find_signposting_linkset(uri)
+            self.assertEqual(uri, metasignposts.context_url)
+            self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/07-http-describedby-citeas-linkset-json/"}, metasignposts.other_contexts)
+            signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/07-http-describedby-citeas-linkset-json/")
+
             self.assertEqual("https://w3id.org/a2a-fair-metrics/07-http-describedby-citeas-linkset-json/", 
                 signposts.citeAs.target)
 
@@ -127,8 +137,10 @@ class TestFindSignpostingLinkset(unittest.TestCase):
     def test_parse_linkset_a2a_08(self):
         PID = "https://w3id.org/a2a-fair-metrics/08-http-describedby-citeas-linkset-txt/"
         for (uri,_) in self._linksets_for_pid(PID):
-            signposts = linkset.find_signposting_linkset(uri)
-            self.assertEqual(uri, signposts.context_url)
+            metasignposts = linkset.find_signposting_linkset(uri)
+            self.assertEqual(uri, metasignposts.context_url)
+            self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/08-http-describedby-citeas-linkset-txt/"}, metasignposts.other_contexts)
+            signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/08-http-describedby-citeas-linkset-txt/")
             
             self.assertEqual("https://w3id.org/a2a-fair-metrics/08-http-describedby-citeas-linkset-txt/", 
                 signposts.citeAs.target)
@@ -153,9 +165,10 @@ class TestFindSignpostingLinkset(unittest.TestCase):
         PID = "https://w3id.org/a2a-fair-metrics/09-http-describedby-citeas-linkset-json-txt/"
         linksets = self._linksets_for_pid(PID)
         for (uri,mediatype) in self._linksets_for_pid(PID):
-            signposts = linkset.find_signposting_linkset(uri, mediatype)
-
-            self.assertEqual(uri, signposts.context_url)
+            metasignposts = linkset.find_signposting_linkset(uri, mediatype)
+            self.assertEqual(uri, metasignposts.context_url)
+            self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/09-http-describedby-citeas-linkset-json-txt/"}, metasignposts.other_contexts)
+            signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/09-http-describedby-citeas-linkset-json-txt/")
             
             self.assertEqual("https://w3id.org/a2a-fair-metrics/09-http-describedby-citeas-linkset-json-txt/", 
                 signposts.citeAs.target)
@@ -179,17 +192,13 @@ class TestFindSignpostingLinkset(unittest.TestCase):
     def test_parse_linkset_a2a_14(self):
         PID = "https://w3id.org/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/"
         for (uri,mediatype) in self._linksets_for_pid(PID):
-            signposts = linkset.find_signposting_linkset(uri, mediatype)
-            if mediatype == MediaType("application/linkset+json"):
-                # Ensure Content-Location was picked up even if "uri" stayed the same
-                self.assertEqual("https://s11.no/2022/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/linkset.json", 
-                    signposts.context_url)
-            elif mediatype == MediaType("application/linkset"):
-                self.assertEqual("https://s11.no/2022/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/linkset.txt", 
-                    signposts.context_url)
-            else:
-                raise Exception("Unknown media type %s" % mediatype)
-            
+            metasignposts = linkset.find_signposting_linkset(uri, mediatype)
+            # Ensure Content-Location was NOT picked up to linkset.json/linkset.txt
+            self.assertEqual("https://s11.no/2022/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/linkset", 
+                metasignposts.context_url)
+            self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/"}, metasignposts.other_contexts)
+            signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/")
+
             self.assertEqual("https://w3id.org/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/", 
                 signposts.citeAs.target)
 
@@ -204,17 +213,19 @@ class TestFindSignpostingLinkset(unittest.TestCase):
                 {i.type for i in signposts.items})
 
             self.assertEqual(3, len(signposts))
-            # Assert context is picked up from anchor="" and not Content-Location
+            # Assert context is picked up from anchor="" and not redirected URL
             for s in signposts:
                 self.assertEqual("https://s11.no/2022/a2a-fair-metrics/14-http-describedby-citeas-linkset-json-txt-conneg/",
                 s.context)
 
     def test_parse_linkset_a2a_27(self):        
-        PID = "https://w3id.org/a2a-fair-metrics/27-http-linkset-json-only//"
+        PID = "https://w3id.org/a2a-fair-metrics/27-http-linkset-json-only/"
         for (uri,_) in self._linksets_for_pid(PID):
-            signposts = linkset.find_signposting_linkset(uri)
-            self.assertEqual("https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/linkset.json", signposts.context_url)
-            
+            metasignposts = linkset.find_signposting_linkset(uri)
+            self.assertEqual("https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/linkset.json", metasignposts.context_url)
+            self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/"}, metasignposts.other_contexts)
+            signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/")
+
             self.assertEqual("https://w3id.org/a2a-fair-metrics/27-http-linkset-json-only/", signposts.citeAs.target)
 
             self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/27-http-linkset-json-only/index.ttl"},
@@ -235,9 +246,11 @@ class TestFindSignpostingLinkset(unittest.TestCase):
     def test_parse_linkset_a2a_28(self):        
         PID = "https://w3id.org/a2a-fair-metrics/28-http-linkset-txt-only/"
         for (uri,_) in self._linksets_for_pid(PID):
-            signposts = linkset.find_signposting_linkset(uri)
-            self.assertEqual("https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/linkset.txt", signposts.context_url)
-            
+            metasignposts = linkset.find_signposting_linkset(uri)
+            self.assertEqual("https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/linkset.txt", metasignposts.context_url)
+            self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/"}, metasignposts.other_contexts)
+            signposts = metasignposts.for_context("https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/")
+
             self.assertEqual("https://w3id.org/a2a-fair-metrics/28-http-linkset-txt-only/", signposts.citeAs.target)
 
             self.assertEqual({"https://s11.no/2022/a2a-fair-metrics/28-http-linkset-txt-only/index.ttl"},